Consolidate duplicate endpoints and reuse service functions
Removed 7 endpoints (62 → 55), eliminated duplicate service logic:

Admin reads → reuse domain services:
- GET /admin/symbols now calls market_service.get_symbols()
- GET /admin/symbols/{id} now calls market_service.get_symbol_by_id()
- GET /admin/pools now calls pool_service.get_all_pools_enriched()
- GET /admin/pools/{id} now calls pool_service.get_pool_detail()
- GET /admin/users/{id} now calls user_service functions
- Deleted get_symbols, get_symbol, get_admin_pools, get_admin_pool,
  get_admin_user from services/admin.py

Dashboard merged:
- GET /admin/dashboard replaces /dashboard/stats + /dashboard/recent-activity
- Merged get_dashboard_stats + get_recent_activity into single get_dashboard()

Removed redundant endpoints:
- GET /pool/public (use /pool + /pool/trades separately)
- GET /pool/liquidity/position (merged into /pool/user)
- GET /market/list_symbols (use /market instead)
- GET /orderbook/orders (use /orderbook/user/orders with optional symbol)

Removed deprecated auth:
- POST /auth/register (use /auth/google)
- POST /auth/login (use /auth/google)
- Deleted register_legacy, login_legacy from services/auth.py

New shared functions:
- market_service.get_symbols(engine_type, is_active, market)
- market_service.get_symbol_by_id(symbol_id)
- pool_service.get_all_pools_enriched()
- pool_service.get_pool_detail(pool_id)
- user_service.get_user_info(user_id)

Closes #54
