+++ backend/routers/admin.py
@@ -1,12 +1,12 @@
 """
-Admin API Routes — thin router, delegates to services/admin.py.
+Admin API Routes — thin router, delegates to domain services.
 
 All endpoints require admin permissions.
 """
 
 from typing import Optional
 
-from fastapi import APIRouter, Depends, Query
+from fastapi import APIRouter, Depends, HTTPException, Query
 
 from backend.core.audit_log import AuditContext, audit_logged, get_audit_context
 from backend.core.auth import require_admin
@@ -24,6 +24,9 @@ from backend.models.admin import (
     UpdateUserStatusRequest,
 )
 from backend.services import admin as admin_service
+from backend.services import market as market_service
+from backend.services import pool as pool_service
+from backend.services import user as user_service
 
 router = APIRouter(prefix="/api/admin", tags=["admin"])
 
@@ -98,42 +101,15 @@ async def delete_symbol(
     return APIResponse(success=True, data=result)
 
 
-# =============================================================================
-# Audit Log Viewer
-# =============================================================================
-
-@router.get("/audit-log", response_model=APIResponse)
-async def get_audit_log(
-    admin_id: Optional[str] = Query(None, description="Filter by admin ID"),
-    action: Optional[str] = Query(None, description="Filter by action type"),
-    target_type: Optional[str] = Query(None, description="Filter by target type"),
-    date_from: Optional[str] = Query(None, description="Start date (ISO format)"),
-    date_to: Optional[str] = Query(None, description="End date (ISO format)"),
-    limit: int = Query(50, ge=1, le=200, description="Page size"),
-    offset: int = Query(0, ge=0, description="Offset"),
-    current_admin: dict = Depends(require_admin),
-):
-    """Query admin audit log with pagination and filters."""
-    data = await admin_service.get_audit_log(
-        admin_id=admin_id, action=action, target_type=target_type,
-        date_from=date_from, date_to=date_to, limit=limit, offset=offset,
-    )
-    return APIResponse(success=True, data=data)
-
-
-# =============================================================================
-# Symbol CRUD (#31)
-# =============================================================================
-
 @router.get("/symbols", response_model=APIResponse)
 async def get_symbols(
     engine_type: Optional[int] = Query(None, description="Filter by engine type: 0=AMM, 1=CLOB"),
-    is_active: Optional[bool] = Query(None, description="Filter by active status"),
+    is_active: Optional[bool] = Query(None, description="Filter by active status (null=all)"),
     market: Optional[str] = Query(None, description="Filter by market type"),
     current_admin: dict = Depends(require_admin),
 ):
-    """Get all symbols with full config. Includes pool info for AMM symbols."""
-    data = await admin_service.get_symbols(engine_type, is_active, market)
+    """Get all symbols with full config. Reuses market_service.get_symbols with is_active=None to include inactive."""
+    data = await market_service.get_symbols(engine_type, is_active, market)
     return APIResponse(success=True, data=data)
 
 
@@ -143,7 +119,7 @@ async def get_symbol(
     current_admin: dict = Depends(require_admin),
 ):
     """Get detailed symbol config + associated pool data."""
-    data = await admin_service.get_symbol(symbol_id)
+    data = await market_service.get_symbol_by_id(symbol_id)
     return APIResponse(success=True, data=data)
 
 
@@ -156,7 +132,7 @@ async def update_symbol(
     current_admin: dict = Depends(require_admin),
     audit: AuditContext = Depends(get_audit_context),
 ):
-    """Update mutable fields of a symbol config. For AMM symbols, also updates pool fee_rate."""
+    """Update mutable fields of a symbol config."""
     data = await admin_service.update_symbol(symbol_id, request, router)
     audit.set(
         target_id=str(symbol_id),
@@ -166,84 +142,13 @@ async def update_symbol(
 
 
 # =============================================================================
-# Platform Settings (#34)
-# =============================================================================
-
-@router.get("/settings", response_model=APIResponse)
-async def get_settings(current_admin: dict = Depends(require_admin)):
-    """Get all platform settings."""
-    data = await admin_service.get_settings()
-    return APIResponse(success=True, data=data)
-
-
-@router.post("/settings/update/{key}", response_model=APIResponse)
-@audit_logged(action="update_setting", target_type="setting")
-async def update_setting(
-    key: str,
-    request: UpdateSettingRequest,
-    current_admin: dict = Depends(require_admin),
-    audit: AuditContext = Depends(get_audit_context),
-):
-    """Update a platform setting."""
-    result = await admin_service.update_setting(key, request.value)
-    audit.set(
-        target_id=key,
-        details={"old": result["old_value"], "new": request.value},
-    )
-    return APIResponse(success=True, data=result["setting"])
-
-
-# =============================================================================
-# Admin Whitelist (#34)
-# =============================================================================
-
-@router.get("/whitelist", response_model=APIResponse)
-async def get_whitelist(current_admin: dict = Depends(require_admin)):
-    """Get all admin whitelist entries."""
-    data = await admin_service.get_whitelist()
-    return APIResponse(success=True, data=data)
-
-
-@router.post("/whitelist", response_model=APIResponse)
-@audit_logged(action="add_whitelist", target_type="whitelist")
-async def add_whitelist(
-    request: AddWhitelistRequest,
-    current_admin: dict = Depends(require_admin),
-    audit: AuditContext = Depends(get_audit_context),
-):
-    """Add an email to admin whitelist."""
-    result = await admin_service.add_whitelist(request.email, request.description)
-    audit.set(
-        target_id=str(result["id"]),
-        details={"email": request.email},
-    )
-    return APIResponse(success=True, data=result)
-
-
-@router.post("/whitelist/remove/{whitelist_id}", response_model=APIResponse)
-@audit_logged(action="remove_whitelist", target_type="whitelist")
-async def remove_whitelist(
-    whitelist_id: int,
-    current_admin: dict = Depends(require_admin),
-    audit: AuditContext = Depends(get_audit_context),
-):
-    """Remove an email from admin whitelist."""
-    result = await admin_service.remove_whitelist(whitelist_id)
-    audit.set(
-        target_id=str(whitelist_id),
-        details={"email": result["email"]},
-    )
-    return APIResponse(success=True, data=result)
-
-
-# =============================================================================
-# Pool Management (#32)
+# Pool Management — reuses pool_service
 # =============================================================================
 
 @router.get("/pools", response_model=APIResponse)
 async def get_admin_pools(current_admin: dict = Depends(require_admin)):
     """Get all AMM pools with enriched data (TVL, price, volume)."""
-    data = await admin_service.get_admin_pools()
+    data = await pool_service.get_all_pools_enriched()
     return APIResponse(success=True, data=data)
 
 
@@ -253,12 +158,12 @@ async def get_admin_pool(
     current_admin: dict = Depends(require_admin),
 ):
     """Get detailed pool info + LP positions + recent swaps."""
-    data = await admin_service.get_admin_pool(pool_id)
+    data = await pool_service.get_pool_detail(pool_id)
     return APIResponse(success=True, data=data)
 
 
 # =============================================================================
-# User Management (#33)
+# User Management — reuses user_service for reads
 # =============================================================================
 
 @router.get("/users", response_model=APIResponse)
@@ -286,8 +191,14 @@ async def get_admin_user(
     current_admin: dict = Depends(require_admin),
 ):
     """Get full user profile + balances + recent trades."""
-    data = await admin_service.get_admin_user(user_id)
-    return APIResponse(success=True, data=data)
+    user = await user_service.get_user_info(user_id)
+    if not user:
+        raise HTTPException(status_code=404, detail=f"User '{user_id}' not found")
+
+    balances = await user_service.get_user_balances(user_id, include_total=True)
+    trades = await user_service.get_user_trades(user_id, limit=50)
+
+    return APIResponse(success=True, data={"user": user, "balances": balances, "trades": trades})
 
 
 @router.post("/users/{user_id}/balance/update", response_model=APIResponse)
@@ -345,21 +256,108 @@ async def reset_user_balances(
 
 
 # =============================================================================
-# Dashboard Stats (#30)
+# Platform Settings
 # =============================================================================
 
-@router.get("/dashboard/stats", response_model=APIResponse)
-async def get_dashboard_stats(current_admin: dict = Depends(require_admin)):
-    """Get aggregated platform statistics."""
-    data = await admin_service.get_dashboard_stats()
+@router.get("/settings", response_model=APIResponse)
+async def get_settings(current_admin: dict = Depends(require_admin)):
+    """Get all platform settings."""
+    data = await admin_service.get_settings()
     return APIResponse(success=True, data=data)
 
 
-@router.get("/dashboard/recent-activity", response_model=APIResponse)
-async def get_recent_activity(
-    period: str = Query("7d", description="Period: 7d or 30d"),
+@router.post("/settings/update/{key}", response_model=APIResponse)
+@audit_logged(action="update_setting", target_type="setting")
+async def update_setting(
+    key: str,
+    request: UpdateSettingRequest,
+    current_admin: dict = Depends(require_admin),
+    audit: AuditContext = Depends(get_audit_context),
+):
+    """Update a platform setting."""
+    result = await admin_service.update_setting(key, request.value)
+    audit.set(
+        target_id=key,
+        details={"old": result["old_value"], "new": request.value},
+    )
+    return APIResponse(success=True, data=result["setting"])
+
+
+# =============================================================================
+# Admin Whitelist
+# =============================================================================
+
+@router.get("/whitelist", response_model=APIResponse)
+async def get_whitelist(current_admin: dict = Depends(require_admin)):
+    """Get all admin whitelist entries."""
+    data = await admin_service.get_whitelist()
+    return APIResponse(success=True, data=data)
+
+
+@router.post("/whitelist", response_model=APIResponse)
+@audit_logged(action="add_whitelist", target_type="whitelist")
+async def add_whitelist(
+    request: AddWhitelistRequest,
+    current_admin: dict = Depends(require_admin),
+    audit: AuditContext = Depends(get_audit_context),
+):
+    """Add an email to admin whitelist."""
+    result = await admin_service.add_whitelist(request.email, request.description)
+    audit.set(
+        target_id=str(result["id"]),
+        details={"email": request.email},
+    )
+    return APIResponse(success=True, data=result)
+
+
+@router.post("/whitelist/remove/{whitelist_id}", response_model=APIResponse)
+@audit_logged(action="remove_whitelist", target_type="whitelist")
+async def remove_whitelist(
+    whitelist_id: int,
+    current_admin: dict = Depends(require_admin),
+    audit: AuditContext = Depends(get_audit_context),
+):
+    """Remove an email from admin whitelist."""
+    result = await admin_service.remove_whitelist(whitelist_id)
+    audit.set(
+        target_id=str(whitelist_id),
+        details={"email": result["email"]},
+    )
+    return APIResponse(success=True, data=result)
+
+
+# =============================================================================
+# Audit Log Viewer
+# =============================================================================
+
+@router.get("/audit-log", response_model=APIResponse)
+async def get_audit_log(
+    admin_id: Optional[str] = Query(None, description="Filter by admin ID"),
+    action: Optional[str] = Query(None, description="Filter by action type"),
+    target_type: Optional[str] = Query(None, description="Filter by target type"),
+    date_from: Optional[str] = Query(None, description="Start date (ISO format)"),
+    date_to: Optional[str] = Query(None, description="End date (ISO format)"),
+    limit: int = Query(50, ge=1, le=200, description="Page size"),
+    offset: int = Query(0, ge=0, description="Offset"),
     current_admin: dict = Depends(require_admin),
 ):
-    """Get daily trade volume and new user counts for charts."""
-    data = await admin_service.get_recent_activity(period)
+    """Query admin audit log with pagination and filters."""
+    data = await admin_service.get_audit_log(
+        admin_id=admin_id, action=action, target_type=target_type,
+        date_from=date_from, date_to=date_to, limit=limit, offset=offset,
+    )
+    return APIResponse(success=True, data=data)
+
+
+# =============================================================================
+# Dashboard — stats + recent activity combined
+# =============================================================================
+
+@router.get("/dashboard", response_model=APIResponse)
+async def get_dashboard(
+    period: str = Query("7d", description="Activity period: 7d or 30d"),
+    current_admin: dict = Depends(require_admin),
+):
+    """Get platform stats + recent activity in one call."""
+    data = await admin_service.get_dashboard(period)
     return APIResponse(success=True, data=data)
